test(tape): added proper support for es6/istanbul/tape for testing
also removed e2e tests for the time being. These are not needed until we
have a full alpha testable product
